pyproject.toml
[tool.poetry]
name = "ayachanbot"
version = "0.1.0"
description = ""
authors = ["myl7 <myl@myl.moe>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.5"
python-dotenv = "^0.17.1"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
